+++ test/HAaveProtocolV2.test.js
@@ -110,6 +110,14 @@ contract('Aave V2', function ([_, user]) {
           from: user,
           value: value,
         });
+
+        const handlerReturn = utils.toBN(
+          getHandlerReturn(receipt, ['uint256'])[0]
+        );
+        expect(
+          await this.aWrappedNativeToken.balanceOf.call(user)
+        ).to.be.bignumber.eq(handlerReturn);
+
         expect(await balanceProxy.get()).to.be.bignumber.zero;
         expect(
           await this.aWrappedNativeToken.balanceOf.call(this.proxy.address)
@@ -134,6 +142,14 @@ contract('Aave V2', function ([_, user]) {
           from: user,
           value: value,
         });
+
+        const handlerReturn = utils.toBN(
+          getHandlerReturn(receipt, ['uint256'])[0]
+        );
+        expect(
+          await this.aWrappedNativeToken.balanceOf.call(user)
+        ).to.be.bignumber.eq(handlerReturn);
+
         expect(await balanceProxy.get()).to.be.bignumber.zero;
         expect(
           await this.aWrappedNativeToken.balanceOf.call(this.proxy.address)
@@ -169,6 +185,14 @@ contract('Aave V2', function ([_, user]) {
           from: user,
           value: ether('0.1'),
         });
+
+        const handlerReturn = utils.toBN(
+          getHandlerReturn(receipt, ['uint256'])[0]
+        );
+        expect(await this.aToken.balanceOf.call(user)).to.be.bignumber.eq(
+          handlerReturn
+        );
+
         expect(await balanceProxy.get()).to.be.bignumber.zero;
         expect(
           await this.aToken.balanceOf.call(this.proxy.address)
@@ -196,6 +220,14 @@ contract('Aave V2', function ([_, user]) {
           from: user,
           value: ether('0.1'),
         });
+
+        const handlerReturn = utils.toBN(
+          getHandlerReturn(receipt, ['uint256'])[0]
+        );
+        expect(await this.aToken.balanceOf.call(user)).to.be.bignumber.eq(
+          handlerReturn
+        );
+
         expect(await balanceProxy.get()).to.be.bignumber.zero;
         expect(
           await this.aToken.balanceOf.call(this.proxy.address)